しょぼカルから過去の番組表を作る
EDCB の EPG は 2026-06 からしか無い。それ以前を番組表で見られるよう、
しょぼカルの放送データを events の形に変換して入れる。

- event_id は放送ID(PID) + 1000000。実 EPG の event_id は 16bit なので
  絶対に衝突せず、合成行だと一目で分かる。
  消すときは DELETE FROM events WHERE event_id >= 1000000
- しょぼカルから EdTime(番組長)・SubTitle(話数タイトル)・Title(作品名)・
  Comment(作品解説) も取るようにし、syobocal_airings / syobocal_titles に列を追加
- Comment は Wiki 記法なので、見出し記号・[[リンク]]・強調を落として素の文章にする
  (CleanWiki)。スタッフ・キャストの記述はそのまま残るので出演者もここに入る
- ジャンルは全てアニメ(nibble_l1=7)
- EpgEvent.EventID を ushort から int へ。1000000 以上を ushort にキャストすると
  値が化けるため

入るのはアニメのみ・しょぼカルに対応表のあるチャンネルのみ。
作品解説は作品単位なので、同じ作品の全話に同じ文章が入る。

+++ Models/EpgEvent.cs
@@ -5,7 +5,11 @@ public class EpgEvent
     public ushort ONID { get; set; }
     public ushort TSID { get; set; }
     public ushort SID { get; set; }
-    public ushort EventID { get; set; }
+    /// <summary>
+    /// 実 EPG の event_id は 16bit だが、しょぼカルから作った合成行は
+    /// 1000000 以上を使うため int で持つ（EpgDbReader.SyntheticEventIdBase）。
+    /// </summary>
+    public int EventID { get; set; }
     public string ServiceName { get; set; } = "";
     public DateTime? StartTime { get; set; }
     public uint? DurationSec { get; set; }
